+++ frontend/app/features/addEvent/controllers/addEventController.js
@@ -1,34 +1,38 @@
 (function() {
     angular.module("em.addEvent").controller("em.addEvent.addEventController", addEventController);
 
-    function addEventController($scope, $rootScope, addEventAPIService) {
-
-      //  $scope.event.date = new Date();
+    function addEventController($scope, uploadService, addEventService) {
+        $scope.event = {
+            isGame: false,
+            report: null
+        };
         $scope.place = {};
-
-        $scope.search = function() {
-                $rootScope.search($scope.event.place)
-                    .then(function(res) { // success
-                            $rootScope.addMarker(res);
-                            $scope.place.name = res.name;
-                            $scope.place.lat = res.geometry.location.lat();
-                            $scope.place.lng = res.geometry.location.lng();
-                        },
-                        function(status) { // error
-                            $scope.apiError = true;
-                            $scope.apiStatus = status;
-                        }
-                    );
-            }
-            // map.init();
-
+        $scope.lookFor = function() {
+            $scope.apiError = false;
+            $scope.search($scope.event.place)
+                .then(
+                    function(res) { // success
+                        $scope.addMarker(res);
+                        $scope.place.name = res.name;
+                        $scope.place.lat = res.geometry.location.lat();
+                        $scope.place.lng = res.geometry.location.lng();
+                    },
+                    function(status) { // error
+                        $scope.apiError = true;
+                        $scope.apiStatus = status;
+                    }
+                );
+        };
+        $scope.upload = function() {
+            uploadService.upload($scope.media).then(function(res) {
+                $scope.uploaded = res.data;
+            }).catch(function(error) {
+                console.log(error);
+            });
+        };
         $scope.add = function() {
-            if ((typeof $scope.event.title !== "undefined") && (typeof $scope.event.date !== "undefined")) {
-                addEventAPIService.add(Object.assign({
-                    report: null
-                }, $scope.event));
-            }
-        }
+            addEventService.addEvent($scope.event);
+        };
     }
-    addEventController.$inject = ["$scope", "$rootScope", "em.addEvent.addEventAPIService"];
+    addEventController.$inject = ["$scope", "em.addEvent.uploadService", "em.addEvent.addEventService"];
 })();
