+++ frontend/app/features/addEvent/directives/file-model.js
@@ -1,5 +1,5 @@
 (function() {
-    angular.module("em.addEvent.v02").directive("fileModel", ["$parse", function($parse) {
+    angular.module("em.addEvent").directive("fileModel", ["$parse", function($parse) {
         return {
             restrict: "A",
             link: function(scope, element, attrs) {
